meta-ibm: add ibm-ac-server MACHINEOVERRIDE
As of today the IBM(witherspoon) layer supports three of IBMs
"Acclerated Compute" class of systems: witherspoon, witherspoon-128 and
swift.

Most of the bbappends are required for all three.  Prior to this patch
this was handled by bbappending without any override.  This is counter
to best practices and pollutes the metadata for layer users using
recipes from meta-witherspoon.

Add a new ibm-ac-server MACHINEOVERRIDE so that bbappends that work on
all of IBM's AC systems can be overridden appropriately without having
to explicitly list every system.

Change-Id: I6687ba51a84fe19e55ae62b624ba0adffec62f60
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>