meta-ibm: witherspoon-128: MACHINEOVERRIDES fixup
The ordering of the values in MACHINEOVERRIDES is wrong:
   MACHINEOVERRIDES="witherspoon-128:witherspoon"

This means that:
   FOO_witherspoon = "bar"

will override:
   FOO_witherspoon-128 = "baz"

which is the opposite of expectation.  The ordering is a function of
config file include order and choice of the predot vs postdot operator.
Use the same include ordering and dot operators as meta-fsl-arm to get
the overrides to appear in MACHINEOVERRIDES in the correct order such
that the expected semantics are provided and:
  FOO_witherspoon-128 = "bar"

will override:
  FOO_witherspoon = "baz"

Change-Id: I170b56a32c96c3d4e7504bf1846ecdec756e49ae
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>