meta-ibm: add tacoma system
The tacoma card is a box elder replacement (witherspoon bmc card) with
an ast2600 on it.

Use aspeed-image-norootfs to work around the absence of a kernel g6 spi
driver.

By default use core-image-minimal as the initramfs.  To add the full
OpenBMC userspace set INITRAMFS_IMAGE="obmc-phosphor-image" in local.conf.

Build instructions:
  TEMPLATECONF=meta-ibm/meta-witherspoon/conf . oe-init-build-env
  MACHINE=witherspoon-tacoma DISTRO=openbmc-openpower bitbake aspeed-image-norootfs

Change-Id: I3e1ca8c7cd47f3f65807f8ebced5d583c39813a8
Signed-off-by: Brad Bishop <bradleyb@fuzziesquirrel.com>